mappings: add raw fields to some external ids
Fields that contain external identifiers need to have a raw variant
to support aggregations and exact searches. This commit adds it to
`arxiv_eprints` and `dois`, but not to other fields (notably, `isbns`
and `report_numbers`) because the data contained in them contains
small typos and irregularities that we need to account for.

Signed-off-by: Jacopo Notarstefano <jacopo.notarstefano@gmail.com>