mappings: add recids from refs to mappings
Since commit 8b0ecbe, when sending records to Elasticsearch we enrich
their JSON refs with the recid they are pointing to. This commit adds
those recids explicitly to the mappings. In fact, Elasticsearch could
easily generate their mappings from the data, but we prefer to state
this constraint explicitly so that in the future we may disallow the
automatic inference of mappings.

The only exception is the `conferences` key in the Authors collection:
this key is currently unused and might be removed in a future version
of the schema.

Signed-off-by: Jacopo Notarstefano <jacopo.notarstefano@gmail.com>