scripts: don't start Celery in recreate_records
When we run that script we must assume that all the other services are
already running, otherwise the script will fail anyway.

Moreover, this removes conflict on what to use as a first argument
passed to the script, as the removed part tried to use it as Celery's
log level, while another part checks to see if it's "--no-populate".

Signed-off-by: Jacopo Notarstefano <jacopo.notarstefano@gmail.com>