tests: use the right fixtures almost everywhere
The `app`, `app_client`, and `api_client` fixtures already provide the
right contexts, so there's no need to explicitly instantiate one.

This is done everywhere except in `orcid`, where the app fixtures are
defined in an obsolete fashion, and `workflows`, where it's possible
they might conflict too much with in-flight PRs.

Signed-off-by: Jacopo Notarstefano <jacopo.notarstefano@gmail.com>