workflows: fix submission number for harvests
The ``submission_number`` key in ``acquisition_source`` is currently
being populated differently between HEPCrawl and form submissions.
In fact, HEPCrawl puts there the Scrapy job id as it can't possibly
know the workflow object id since it will be generated later. This
workflow step makes sure that they will both contain the workflow id
as we'd might want to link from the record to the Holding Pen entry
from which it came from.

Signed-off-by: Iuliana Voinea <iuliana.voinea@student.manchester.ac.uk>
Co-authored-by: Jacopo Notarstefano <jacopo.notarstefano@cern.ch>