config: don't register invenio-collection signals
Because we are instantiating `invenio-collection` two times, we can't
have it register its signals on instantiation, as we end up with the
same signal registered twice and no way of unregistering one of them.

Therefore, we don't need to unregister them during migration, as they
were never registered in the first place.

Signed-off-by: Jacopo Notarstefano <jacopo.notarstefano@gmail.com>