setup: add theme extension to the API application
Adds the ``inspire-theme`` extension to the API application so that we
can render Jinja templates in API responses.

This required modifying the existing error handlers so that they return
a JSON response instead of trying to render an error template, as that
would fail because the API application doesn't (and shouldn't) know
about ``invenio-assets``.

In order to do that, injects the ``API_RESTFUL`` configuration variable
only in the API application when the configuration is instantiated.

Signed-off-by: Szymon Łopaciuk <szymon.lopaciuk@cern.ch>