global: handle files between records and workflows
Adds the ``files`` property to ``InspireRecord`` by inheriting from the
``Record`` of ``invenio_records_files``. This allows us to attach plots
and documents to records (closes #2551 and closes #2552).

Also modifies those workflows tasks that deal with records so that they
use builder methods to populate the ``documents`` and ``figures`` keys,
and those will then be converted by DoJSON to the ``FFT`` fields that
Legacy understands.

Signed-off-by: Zacharias Zacharodimos <zacharias.zacharodimos@cern.ch>