setup: bump invenio-records and unpin SQLAlchemy
Fixes the build by removing the unsatisfiable pin on SQLAlchemy that
came with the latest release of Invenio-Records. We preferred this
solution to a pin on the latter because we want to profit from the
added JSONB support.

On the other hand, changing the type means that we can no longer use
the ``json_array_elements`` function. Since we are not yet ready to
migrate all our production data we preferred to add an explicit cast.

Finally, the newest version of Invenio-Records also came with an
incompatible API change in the receivers, also fixed by this commit.

Signed-off-by: Jacopo Notarstefano <jacopo.notarstefano@gmail.com>