Workflows: normalize journal title
I added a step to the workflows that takes the journal title of every journal
that is found in `publication_info` and then normalizes it by performing a
scan through the Refextract Journal KB File. Moreover, it then queries the
database and adds/updates the journal's `self` field value (reference) to the `journal_record`
field, if found.

Signed-off-by: Iuliana Voinea <iuliana.voinea@student.manchester.ac.uk>