mappings: add raw field to isbns.value
The key ``value`` in ``isbns`` needs a raw variant as we want to
search for its exact value while matching references.

Note however that we lowercase it rather than keeping it unchanged,
as the name could imply. This might be changed in the future, after
we decide on a naming convention for fields.

Signed-off-by: Jacopo Notarstefano <jacopo.notarstefano@gmail.com>