receivers: populate bookautocomplete in Python
The previous implementation was trying to mix the ``copy_to`` directive
in the Elasticsearch mappings and some Python to copy all needed fields
in ``bookautocomplete``, but this couldn't possibly work as at indexing
time the content of ``bookautocomplete`` is not known.

Therefore this commit rewrites all that logic in Python only.

Signed-off-by: Iuliana Voinea <iuliana.voinea@student.manchester.ac.uk>