config: make journal suggester return 10 hits
Previously the journal suggester only displayed 5 results, and
not necessarily the most relevant ones. This is due to the fact
that ES completion types are implemented in such a way that they
don't support relevance-based sorting. Therefore this commit tries
to improve relevance by showing more results, i.e. 10 instead of 5.

Signed-off-by: Iuliana Voinea <iuliana.voinea@student.manchester.ac.uk>