Replace `confereneautocomplete` with `conference_suggest`
I replaced the old `conferenceautocomplete` with `conference_suggest`. Also,
there was a bug in the JavaScript which called the `conference_suggest` endpoint
every time the submission form was laoded, which I also hopefully fixed.

Lastly, I removed some commented/unused autocompletion functions from the forms
module and added `TODO`. comments which mention the fact that we need a consistent
way of performing autocompletion (with ES compeltion suggeters), instead of using
the old way this was implemented in `INSPIREForm`.

Signed-off-by: Iuliana Voinea <iuliana.voinea@student.manchester.ac.uk>