records: populate full_name_unicode_normalized
Add ``authors.full_name_unicode_normalized`` to the HEP mapping in order
to support exact author queries. This field is populated by a receiver
so as to ensure that the same work is being done at indexing and at
query time: in both we use the NFKC unicode normalization form from
unicodedata while also lowercasing the value. Note also that we assume
that ``authors.full_name`` has already been normalized by the
``normalize_name`` util from ``inspire-utils``.

See inspirehep/inspire-query-parser#63 for the rational behind the
choice of the normalization form, inspirehep/inspire-query-parser#64
and inspirehep/inspire-query-parser#84 for the implementation.

Signed-off-by: Iuliana Voinea <iuliana.voinea@student.manchester.ac.uk>