workflow: integrate fuzzy matcher in the `article` workflow
Changed the step in the workflow that checks in the ingested article is an update, by
performing a fuzzy match. If there is at least a match, the workflow halts, waiting
for human matching confirmation.

All the matches are now stored in `obj.extra_data.matches`, which is a doctionary containing
`exact` matches, `fuzzy` matches, and eventually the `approved` fuzzy match.

A new action called `MatchApproval` has been added to resume the workflow when is halted for
matching approval.

Signed-off-by: Antonio Cesarano <cesarano2607@gmail.com>