gitignore: ignore also symlink to the var folder
We currently ignore the var dir. This works well if the project folder is inside the virtualenvironment. But if the project folder is somewhere else, then you might want to add a symlink to the var folder. Thus, ignoring both the var folder and var symlink is the best solution.