migrator: don't create another version on ES
Avoid creating another version on Elasticsearch when adding the
citation counts because this creates an inconsistency between
the version in the DB and the version on ES, which means that
an exception will be raised by the indexing receiver since ES
will reject an update if its version is lower than the current.

The specific mechanism used (``_version_type: force``) has been
deprecated in ES@2 and removed in ES@5, so we will need to find
another solution when we will upgrade Elasticsearch.

Signed-off-by: Jacopo Notarstefano <jacopo.notarstefano@gmail.com>
Co-authored-by: Chris Aslanoglou <chris.aslanoglou@gmail.com>