mappings: add page_artid to publication_info
Add new `artid_and_page_start` field to enable easier queries based on
`artid` and `page_start`. Because the data is ambiguous, the 2 fields
are sometimes populated with the same value, different values or one of
them is not populated at all. By using `copy_to` to copy the *values*
stored in each one of them to `artid_and_page_start`, we can now query
a single field instead of all the possible combinations of the other 2
fields.

Signed-off-by: Iuliana Voinea <iuliana.voinea@student.manchester.ac.uk>