workflows: add reference extraction from raw_refs
This commit adds reference extraction from `raw_refs` to the
`refextract` workflow step. When the workflow metadata contains a
`references` field (as might happen when harvesting from journals),
`refextract` is not run on the PDF, but instead on the elements of
`references` that have a `raw_refs` but no `reference`, that is they have
a raw reference but no structured reference.

Signed-off-by: Micha Moskovic <michamos@gmail.com>