mappings: make authors nested
We record that an author of recid X has claimed a paper of recid Y by
setting ``curated_relation`` to ``True`` in the author signature of Y
that also contains a record reference to X.

In order to later retrieve this information we need to nest this field,
otherwise we would get spurious matches across different signatures.
See also c20f672 for a similar argument for ``publication_info``s.

Signed-off-by: Jacopo Notarstefano <jacopo.notarstefano@gmail.com>