e2e: add first e2e test for arxiv harvest
This commit adds a new test suite called e2e in our docker
environment, which spins up all the containers needed for e2e tests.
A `fake-arxiv-service` class is added and instantiated in a new container
in the e2e tests.
Furthermore, a new end-to-end test for harvesting a core article from arxiv
has beeen added.
Legacy, RT, and ORCID services are not mocked and bypassed by the
`PRODUCTION-MODE` feature flag.

Signed-off-by: Antonio Cesarano <cesarano2607@gmail.com>