workflows: integrate reference_matcher in labs during refextract
Currently, Labs is using the Legacy Reference matcher. This commit upgrades the the reference matcher by integrating the new Labs reference matcher, which relies on inspire_matcher and ElasticSearch. It also takes into account citations for data. This upgrades the refextract in the workflows, by retaining references either from text or pdf, depnding not on the qunatity but on the quality of the references i.e. how many of those match to other records in INSPIRE. The source with the most matches would be retained and added to the wrokflow object.

- Signed- off-by: Salman Maqbool <salman.maqbool@cern.ch>