workflows: added integration tests for reference matcher
The integration test inserts a cited record into the database and indexes it in ElasticSearch. Another record is then run through the workflow which contains a reference to the cited record. refextract then tries to match the references extracted from the 2nd document. If the reference for the 1st record is matched succesfully, the test passes.

Signed-off-by: Salman Maqbool <salman.maqbool@cern.ch>