utils: rename modules.cache.utils -> utils.lock
Also renames the context manager to `distributed_lock` as the caller is
not required to use Redis at all. On the contrary, using a lock to
operate on Redis is probably not what should be done, as Redis has
built-in mechanisms to ensure isolation. For this reason, the
`StrictRedis` object is no longer passed inside the `with` block.

Signed-off-by: Micha Moskovic <michamos@gmail.com>