merger: introduce the use of root
Integrate the rootful merger in the article workflow for updates. Previously,
the merger was rootless. The manual merge remains rootless, as well as merging
submissions. In the case of the manual merger, the only change is the logic
behind selecting, saving and assigning the roots to the new merged record.

Signed-off-by: Iuliana Voinea <iuliana.voinea@student.manchester.ac.uk>