workflows: account for ambigious/errata citations
In addition to refactoring the reference matcher code, this commit imposes
conditions on the reference matcher to take a conservative approach when
matching references. A match is confirmed only if there's only 1 match OR
one of the matches is also the same as the previous match> The second case
is also typical of errata citatons.

Signed-off-by: Salman Maqbool <salman.maqbool@cern.ch>