utils: make is_pdf_link accept long lines
The previous (i.e. third) iteration of `is_pdf_link` didn't work when
all the content was on a single line, as it was then reading too much
resulting in the connection being killed.
This is avoided by reading a fixed number of bytes, instead of a fixed
number of lines.

Signed-off-by: Micha Moskovic <michamos@gmail.com>