907 workflows: make HP hep mapping non-dynamic
This disables the default behaviour of the dynamic mapping creation,
which can cause conflicts when two records with conflicting types,
e.g. in extra_data, are indexed. Instead, this will only enable indexing
on fields specified explicitly and will ignore the rest.

Also fixed a small mistake in docs related to the remapping.

Signed-off-by: Szymon Łopaciuk <szymon@lopaciuk.eu>
Actions-after-deployment: remap, reindex