workflows: put the refextract timeout on subtasks
Increasing the timeout to 10 min in
92d3166d84652520cf294aef430ae9c8296ea9de didn't solve the problem. This
might be because when there are lots of references, also reference
matching takes a long time, exceeding the allowed time. As we really
want to protect from refextract going wild and never finishing, rather
than ensuring the task runs in a fixed amount of time, it makes sense
not to include reference matching in the timeout.

Signed-off-by: Micha Moskovic <michamos@gmail.com>