records: move logic from wrapper to serializer
Move the serialization logic from the `LiteratureRecord` wrapper class
to the actual records serializers. The schemas for the UI serializers
endpoints are implemeted by using the `marshmallow` library.

New schemas have been created for `authors` and `supervisors` that need
to be filtered before being serialized and returned to the UI.

Signed-off-by: Iuliana Voinea <iulianavoinea96@gmail.com>