1313 orcid: generate random email when email is blank in import_legacy_orcid_tokens
Legacy might send (orcid, token, email, name) with an empty string as email or name (because this info was not found in the PERSON table nor in HepNames, yet there is a valid orcid token).

Labs creates a User and email is a mandatory and unique field.
To solve this issue, we generate a random string as email (in case the email is empty) following a pattern.