citation count: index new citation's records - INSPIR-1127
Changes in the current receiver models_committed to trigger a celery task,
which computes the references diff between the current and previous
version of the record and indexes the cited records.
However, before computing the references diff, the task checks that the
current version of the record in the DB is the one it expects to find
(passed before by the receiver). If it is not, it retries itself with exponential time.
A new celery queue called bulk_index is used for reindexing the cited records.

Signed-off-by: Antonio Cesarano <cesarano2607@gmail.com>